requirements.txt
flask
flask-cors
uvicorn
fastapi
Markdown
numpy
uv
requests
beautifulsoup4
pandas
pillow
pdfplumber
scikit-learn
dotenv